Extracted logic to convert test input to CSV file
